merge

+++ TestSDL/AffichageGeneral.c
@@ -33,7 +33,7 @@ int mainFenetre(Jeu * jeu)
 			cleanUp(&pWindow, &pRenderer, &pInput);
 			return -1;
 		}
-		   
+
 		//Initialisation worms
 		worms1 = createWorms("Jacouille");
 		if (worms1 == NULL)
@@ -476,18 +476,18 @@ void initCameras(SDL_Renderer * pRenderer, Terrain * map, SDL_Rect * camera, Wor
 	int w = 0, h = 0, wW = 0, hW = 0;
 	SDL_GetRendererOutputSize(pRenderer, &wW, &hW);
 	if (worms == NULL){
-	w = map->imageMapSurface->w;
-	h = map->imageMapSurface->h;
-	camera->x = 0;
-	camera->y = 0;
+		w = map->imageMapSurface->w;
+		h = map->imageMapSurface->h;
+		camera->x = 0;
+		camera->y = 0;
 		if (h > hW || w > wW){
-		camera->h = h;
-		camera->w = (int)(camera->h * ((float)wW / (float)hW));
-	}
+			camera->h = h;
+			camera->w = (int)(camera->h * ((float)wW / (float)hW));
+		}
 	}
 	else{
 		camera->h = worms->wormsSurface->h * 10;
-		camera->w = camera->h * ((float)wW / (float)hW);
+		camera->w = camera->h * (int)((float)wW / (float)hW);
 		camera->x = worms->xAbs - (camera->w / 2);
 		camera->y = worms->yAbs - (camera->h / 2);
 		if (camera->x < 0)camera->x = 0;
@@ -546,10 +546,10 @@ void zoomIn(SDL_Renderer * pRenderer, SDL_Texture * pTexture, SDL_Rect * camera,
 	camera->h = camera->h - 10;
 	camera->w = (int)(camera->h * ((float)wW / (float)hW));// keep the ratio depending of the size of the window!!!!!
 
-	x = (pInput->cursor.now.x / (float)wW) - (1 / 2.0);
-	y = (pInput->cursor.now.y / (float)hW) - (1 / 2.0);
-	camera->x += x * 30;
-	camera->y += y * 30;
+	x = (float)(pInput->cursor.now.x / (float)wW) - (float) (1 / 2);
+	y = (float)(pInput->cursor.now.y / (float)hW) - (float) (1 / 2);
+	camera->x += (int) x * 30;
+	camera->y += (int) y * 30;
 
 	if (camera->x < 0)camera->x = 0;
 	if (camera->y < 0)camera->y = 0;
@@ -664,12 +664,12 @@ int createGlobalTexture(SDL_Surface* pSurfaceMap, SDL_Texture** pTexture, SDL_Re
 		{
 			pixelRead = ReadPixel(pSurfaceMap, x - pSurfaceMap->clip_rect.x, y - pSurfaceMap->clip_rect.y);
 			SDL_GetRGBA(pixelRead, pSurfaceMap->format, &r, &g, &b, &a);
-				if (a < 150)
-				{
+			if (a < 150)
+			{
 				pixelWrite[x + y *  pSurfaceMap->w] = pixelTransparent;
-				}
-				else
-				{
+			}
+			else
+			{
 				pixelWrite[x + y *  pSurfaceMap->w] = pixelRead;
 			}
 		}
